[CHE-8] Move OrderItem into its own file.
